REF: implement logical and comparison array ops
